1.更正大师头像 2.增加百度小程序登入

+++ controllers/api/common.go
@@ -8,7 +8,7 @@ import (
 
 var (
 	Token string
-	Appid int
+	Appid int64
 )
 
 type CommonController struct {
